refactor: refactored & modularized components to add additional quiz types
