+++ src/components/anime/AnimeAPI.tsx
@@ -1,21 +1,4 @@
-import { shuffleArray } from './utils';
-
-export type Question = {
-  category: string;
-  correct_answer: string;
-  difficulty: string;
-  incorrect_answers: string[];
-  question: string;
-  type: string;
-};
-
-export type QuestionState = Question & { answers: string[] };
-
-export enum Difficulty {
-  EASY = 'easy',
-  MEDIUM = 'medium',
-  HARD = 'hard',
-}
+import { Difficulty, Question, shuffleArray } from '../../utils/utils';
 
 export const fetchQuizQuestions = async (
   amount: number,
