+++ src/index.tsx
@@ -2,7 +2,8 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import './index.css';
-import App from './App';
+import Anime from './components/anime/Anime';
+import Comics from './components/comics/Comics';
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
@@ -11,7 +12,8 @@ root.render(
   <React.StrictMode>
     <BrowserRouter>
       <Routes>
-        <Route path="/anime" element={<App />} />
+        <Route path="/anime" element={<Anime />} />
+        <Route path="/comics" element={<Comics />} />
       </Routes>
     </BrowserRouter>
   </React.StrictMode>
